chore: prepare for multiple crates
Prepare to split the project into to crates, since the current way of
handling base fields for properties is unmaintainable. A new proc macro
crate will be added and the project will convert to a Cargo workspace,
instead of the current standalone project structure.

Update the CI, README and `Cargo.toml` files in preparation for it.

Signed-off-by: Filip Dutescu <filip@hucksy.dev>