feat(components): add the modal component
Implement the [Bulma modal components][bd] as a [Yew][yew]
component. Document them and their properties and add an example
showcasing how it's used. Update the module tree to include the
modal module and document the new additions to the crate.

Add a macro for making it easier to set all base component attributes.

[bd]: https://bulma.io/documentation/components/modal/
[yew]: https://yew.rs

Fixes: #36
Signed-off-by: Filip Dutescu <filip@hucksy.dev>