feat: add a base component
Add a `BaseComponent` which is tasked with adding all the base HTML
attributes that each component should expose. It takes a tag, children
and the base attributes. It also defines a new property, `attrs` which
contains the extra attributes that a component should have defined. This
includes non-standard ones, such as `data-*`.

Add some [ARIA attributes][aria] to the base properties. Modify the
`base_component_properties` to also implement `From<T>` for
`BaseComponentProperties`, in order to make it easy to pass the default
properties down to the `BaseComponent`. This is also implemented for a
reference to `T`.

This also updates all components and their examples.

[aria]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes

Fixes: #14
Signed-off-by: Filip Dutescu <filip@hucksy.dev>