feat(components): add the panel component
Implement the [Bulma panel components][bd] as a [Yew][yew] component.
Document them and their properties and add an example showcasing how
how it's used. Update the module tree to include the panel module and
document the new additions to the crate.

Update the `class` base component property to not be an `Option<T>` in
order to have a more intuitive API.

[bd]: https://bulma.io/documentation/components/panel/
[yew]: https://yew.rs

Fixes: #39
Signed-off-by: Filip Dutescu <filip@hucksy.dev>